feat: add day 10

+++ src/day_13.rs
@@ -1,3 +1,3 @@
-pub fn solve(input:String) -> (u32, u32) {
+pub fn solve(input:String) -> (u64, u64) {
   return (0, 0);
 }
